cheri: Fix compile in driver/gpu/
- ioctl issues
- Standard uintptr_t vs unsigned long issues.
- Some UAPI structures need to change.

+++ drivers/gpu/drm/amd/pm/powerplay/hwmgr/pp_psm.c
@@ -94,7 +94,7 @@ int psm_init_power_state_table(struct pp_hwmgr *hwmgr)
 
 		if (state->classification.flags & PP_StateClassificationFlag_Uvd)
 			hwmgr->uvd_ps = state;
-		state = (struct pp_power_state *)((unsigned long)state + size);
+		state = (struct pp_power_state *)((uintptr_t) state + size);
 	}
 
 	return 0;
@@ -133,7 +133,7 @@ static int psm_get_ui_state(struct pp_hwmgr *hwmgr,
 			*state_id = state->id;
 			return 0;
 		}
-		state = (struct pp_power_state *)((unsigned long)state + hwmgr->ps_size);
+		state = (struct pp_power_state *)((uintptr_t) state + hwmgr->ps_size);
 	}
 	return -EINVAL;
 }
@@ -154,7 +154,7 @@ static int psm_get_state_by_classification(struct pp_hwmgr *hwmgr,
 			*state_id = state->id;
 			return 0;
 		}
-		state = (struct pp_power_state *)((unsigned long)state + hwmgr->ps_size);
+		state = (struct pp_power_state *)((uintptr_t) state + hwmgr->ps_size);
 	}
 	return -EINVAL;
 }
@@ -174,7 +174,7 @@ static int psm_set_states(struct pp_hwmgr *hwmgr, unsigned long state_id)
 			memcpy(hwmgr->request_ps, state, hwmgr->ps_size);
 			return 0;
 		}
-		state = (struct pp_power_state *)((unsigned long)state + hwmgr->ps_size);
+		state = (struct pp_power_state *)((uintptr_t) state + hwmgr->ps_size);
 	}
 	return -EINVAL;
 }
